third_party/WebKit/LayoutTests/fast/shapes/parsing/parsing-shape-property-aliases.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script src="parsing-test-utils.js"></script>
<script>
description('Test that CSS Shapes properties with the -webkit prefix alias the unprefixed properties.');

// This test is only necessary until the prefixed CSS properties have been removed

function getAliasCSSValue(alias, property, value)
{
    var element = document.createElement("div");
    element.style.setProperty(alias, value);
    return element.style.getPropertyValue(property);
}

function testPropertyAlias(alias, property, value)
{
    shouldBeEqualToString("getAliasCSSValue('" + alias + "','" + property + "','" + value +"')", value);
}

applyToEachArglist(
    testPropertyAlias,
    [// [alias, property, value]
     ["shape-outside", "shape-outside", "circle(0px at 0% 0%)"],
     ["shape-margin", "shape-margin", "10px"],
     ["shape-image-threshold", "shape-image-threshold", "0.5"]
    ]
);
</script>
</body>
</html>
